+++ native/gpu/scripts/make.mjs
@@ -4,7 +4,7 @@ import { execFileSync } from 'child_process'
 import C from './util/common.js'
 
 console.log("build in", C.dir.build)
-execFileSync(C.dir.ninja, ['-C', C.dir.build, 'dawn.node', '-v'], {
+execFileSync(C.dir.ninja, ['-C', C.dir.build, 'pixi_native_gpu.node', '-v'], {
 	stdio: 'inherit',
 	env: {
 		...process.env,
@@ -16,8 +16,8 @@ console.log("copy to", C.dir.dist)
 await Fs.promises.rm(C.dir.dist, { recursive: true }).catch(() => {})
 await Fs.promises.mkdir(C.dir.dist, { recursive: true })
 await Fs.promises.cp(
-	Path.join(C.dir.build, 'dawn.node'),
-	Path.join(C.dir.dist, 'dawn.node'),
+	Path.join(C.dir.build, 'pixi_native_gpu.node'),
+	Path.join(C.dir.dist, 'pixi_native_gpu.node'),
 )
 
 if (C.platform === 'win32') {
@@ -29,5 +29,5 @@ if (C.platform === 'win32') {
 
 // Strip binaries on linux
 if (C.platform === 'linux') {
-	execFileSync('strip', ['-s', Path.join(C.dir.dist, 'dawn.node')])
+	execFileSync('strip', ['-s', Path.join(C.dir.dist, 'pixi_native_gpu.node')])
 }